requirements.txt
alembic==0.8.9
asn1crypto==0.24.0
Authlib==0.8
bcrypt==3.1.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.2
cryptography==2.2.2
enum34==1.1.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT-Simple==0.0.3
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.1
functools32==3.2.3.post2
gevent==1.3.4
gevent-websocket==0.10.1
greenlet==0.4.13
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.6
MarkupSafe==1.0
netifaces==0.10.7
pkg-resources==0.0.0
postgres==2.2.1
psycopg2==2.6.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-editor==1.0.3
pyzmq==17.0.0
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
socketIO-client==0.7.2
SQLAlchemy==1.2.9
ujson==1.35
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
ws4py==0.5.1
wsaccel==0.6.2
